fix: Fixed test_get_episodes_action
