+++ src/gpodder/auth/authentication.rs
@@ -162,8 +162,7 @@ mod tests {
     use crate::test_utils::test::{ContainerCommands, POSTGRES_CHANNEL};
     use crate::utils::test_builder::device_test_builder::tests::DevicePostTestDataBuilder;
     use crate::utils::test_builder::user_test_builder::tests::UserTestDataBuilder;
-    use base64::engine::general_purpose;
-    use base64::Engine;
+    use crate::utils::auth::tests::create_auth_gpodder;
 
     #[tokio::test]
     #[serial]
@@ -175,21 +174,9 @@ mod tests {
             .unwrap();
         let mut user = UserTestDataBuilder::new().build();
         user.insert_user().expect("TODO: panic message");
-        let encoded_auth =
-            general_purpose::STANDARD.encode(format!("{}:{}", user.username, "password"));
-        server.clear_headers();
-        server.add_header("Authorization", format!("Basic {}", encoded_auth));
 
-        let response = server
-            .post(&format!("/api/2/auth/{}/login.json", user.username))
-            .await;
+        create_auth_gpodder(&mut server, &user).await;
 
-        assert!(response.status_code().is_success());
-        assert!(response.cookies().get("sessionid").is_some());
-
-        // get devices
-        let cookie_binding = response.cookies();
-        server.add_cookie(cookie_binding.get("sessionid").unwrap().clone());
         let response = server
             .get(&format!("/api/2/devices/{}", user.username))
             .await;
