+++ src/utils/mod.rs
@@ -12,3 +12,4 @@ pub mod reqwest_client;
 pub(crate) mod rss_feed_parser;
 pub mod test_builder;
 pub mod time;
+pub mod auth;
